Introduce `Codec` for stripping and wrapping in SSH message frames
